requirements.txt
aiohttp
asyncpg
python-dotenv
apscheduler
beautifulsoup4
